requirements.txt
gym==0.9.4
keras==2.0.8
keras-rl==0.4.0
pyyaml==3.12